pyproject.toml
[tool.poetry]
name = "acad-gpt"
version = "0.0.0"
description = "A service for distilling papers, github repos, youtube tutorials and much more using the power of LLMs."

authors = [
    "Shahrukh Khan <sk28671@gmail.com>",
    "Navdeeppal Singh <npsingh0181@gmail.com>"
]
readme = "README.md"
packages = [{include = "acad_gpt"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.0"
pytest = "^7.2.2"
numpy = "^1.24.2"
tqdm = "^4.65.0"
requests = "^2.28.2"
transformers = "^4.26.1"
redis = "^4.5.1"
openai = "^0.27.2"
langchain = "^0.0.113"
python-dotenv = "^1.0.0"
scipdf-mirror = "^0.1.dev0"
bs4 = "^0.0.1"
fastapi = "^0.95.1"
uvicorn = "^0.21.0"
python-multipart = "^0.0.5"
pynvml = "^11.0"
psutil = "^5.9"
huggingface-hub = "^0.14.1"
elasticsearch = "^7.0.0"
discord = "^2.2.2"
validators = "^0.20.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
filter_files = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py310']

[tool.ruff]
 line-length = 120
 select = [
     "E", "W",  # see: https://pypi.org/project/pycodestyle
     "F",  # see: https://pypi.org/project/pyflakes
 ]
 ignore = [
    "E501",
    "E731",
    "E741",  # Ambiguous variable name: ...
    "E999",  # SyntaxError: invalid syntax. Got unexpected token Newline
 ]
 exclude = [
    ".eggs",
    ".git",
    ".ruff_cache",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "docs"
 ]
 ignore-init-module-imports = true
